metachat-aim/requirements.txt
adjusttext==1.3.0 ; python_version >= "3.12" and python_version < "4.0"
aiofiles==24.1.0 ; python_version >= "3.12" and python_version < "4.0"
aiohappyeyeballs==2.4.3 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.11.8 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
aiosqlite==0.20.0 ; python_version >= "3.12" and python_version < "4.0"
alembic==1.14.0 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anthropic==0.39.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.6.2.post1 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
contourpy==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.12" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
eval-type-backport==0.2.2 ; python_version >= "3.12" and python_version < "4.0"
filelock==3.16.1 ; python_version >= "3.12" and python_version < "4.0"
fonttools==4.55.3 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version == "3.12" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
jiter==0.7.1 ; python_version >= "3.12" and python_version < "4.0"
joblib==1.4.2 ; python_version >= "3.12" and python_version < "4.0"
kaleido==0.2.1 ; python_version >= "3.12" and python_version < "4.0"
kiwisolver==1.4.8 ; python_version >= "3.12" and python_version < "4.0"
llamaapi==0.1.36 ; python_version >= "3.12" and python_version < "4.0"
llvmlite==0.44.0 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.6 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
matplotlib==3.10.0 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
mpmath==1.3.0 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.1.0 ; python_version >= "3.12" and python_version < "4.0"
nest-asyncio==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
numba==0.61.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.1.3 ; python_version >= "3.12" and python_version < "4.0"
openai==1.54.4 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.12" and python_version < "4.0"
propcache==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
pyarrow==18.1.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.27.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
pynndescent==0.5.13 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.2.1 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rich==13.9.4 ; python_version >= "3.12" and python_version < "4.0"
scikit-learn==1.6.1 ; python_version >= "3.12" and python_version < "4.0"
scipy==1.14.1 ; python_version >= "3.12" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.35 ; python_version >= "3.12" and python_version < "4.0"
sympy==1.13.3 ; python_version >= "3.12" and python_version < "4.0"
tabulate==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
threadpoolctl==3.5.0 ; python_version >= "3.12" and python_version < "4.0"
together==1.3.11 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.67.0 ; python_version >= "3.12" and python_version < "4.0"
typer==0.15.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.12" and python_version < "4.0"
umap-learn==0.5.7 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.18.0 ; python_version >= "3.12" and python_version < "4.0"
